318 - Adds useWithOrchestrator hook and extracts reusable WFOJsonCodeBlock
