+++ packages/orchestrator-ui-components/src/components/WFOJsonCodeBlock/styles.ts
@@ -1,0 +1,13 @@
+import { css } from '@emotion/react';
+import { EuiThemeComputed } from '@elastic/eui';
+
+export const getStyles = (theme: EuiThemeComputed) => {
+    const euiCodeBlockStyle = css({
+        marginTop: 10,
+        borderRadius: theme.border.radius.medium,
+    });
+
+    return {
+        euiCodeBlockStyle,
+    };
+};
