--- packages/orchestrator-ui-components/src/components/WFOWorkflowSteps/WFOStepsRawJson/index.ts
@@ -1,1 +1,0 @@
-export * from './WFOStepsRawJson';
