+++ packages/orchestrator-ui-components/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './useSubscriptionProcesses';
 export * from './useDataDisplayParams';
 export * from './useToastMessage';
 export * from './useStoredTableConfig';
+export * from './useWithOrchestratorTheme';
