+++ packages/orchestrator-ui-components/src/hooks/useWithOrchestratorTheme.ts
@@ -1,0 +1,9 @@
+import { EuiThemeComputed } from '@elastic/eui';
+import { useOrchestratorTheme } from './useOrchestratorTheme';
+
+export function useWithOrchestratorTheme<T>(
+    getStylesFunction: (theme: EuiThemeComputed) => T,
+) {
+    const { theme } = useOrchestratorTheme();
+    return getStylesFunction(theme);
+}
